Add second op with dependency.
Your code was observing `finished` and `executing`, but the real issue is that in concurrent/asynchronous operations, for dependencies to work, you must post `isFinished` and `isExecuting` notifications. BTW, to notice this key difference, you have to make the operation complete asynchronously (your original implementation made it complete synchronously, and therefore the failure to do the appropriate KVN was not relevant).

I pulled out your observing `finished` and `executing`, as that's immaterial to `NSOperation` dependencies.